&&
ttps://lintr.r-lib.org/reference/vector_logic_linter.html
Enforce usage of scalar logical operators in conditional statements
Source: R/vector_logic_linter.R
Usage of & in conditional statements is error-prone and inefficient. condition in if (condition) expr must always be of length 1, in which case && is to be preferred. Ditto for | vs. ||.